# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Advanced web panel for managing Xray-core servers. Supports VMESS, VLESS, Trojan, Shadowsocks, Hysteria, WireGuard, and more.",
        "title": "3X-UI Panel API",
        "termsOfService": "https://github.com/MHSanaei/3x-ui",
        "contact": {
            "name": "MHSanaei",
            "url": "https://github.com/MHSanaei/3x-ui"
        },
        "license": {
            "name": "GNU GPL v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "3.0.1"
    },
    "host": "localhost:2053",
    "basePath": "/",
    "paths": {
...
        "/panel/api/server/cpuHistory/{bucket}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Legacy: aggregated CPU history. Use /history/cpu/{bucket} instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get CPU history (legacy)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bucket size in seconds (2, 30, 60, 120, 180, 300)",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getConfigJson": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the assembled Xray configuration that is currently running on this host.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get Xray config JSON",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getDb": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exports the SQLite database file as an attachment. Requires password verification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Export database",
                "parameters": [
                    {
                        "description": "Password for verification",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.getDbPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Database file download",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewEchCert": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new Encrypted Client Hello (ECH) keypair.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate ECH keypair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SNI for ECH certificate",
                        "name": "sni",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewUUID": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a fresh UUID v4. Convenience helper for client IDs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate UUID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewVlessEnc": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new VLESS encryption keypair.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate VLESS encryption key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewX25519Cert": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new X25519 keypair for Reality TLS. Returns public and private keys.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate X25519 keypair",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewmldsa65": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new ML-DSA-65 post-quantum signature keypair.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate ML-DSA-65 keypair",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getNewmlkem768": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new ML-KEM-768 post-quantum KEM keypair.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Generate ML-KEM-768 keypair",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getPanelUpdateInfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Checks whether a newer 3x-ui release is available on GitHub.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Check panel updates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/api/server/history/{metric}/{bucket}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns up to 60 buckets of aggregated time-series data for a single metric (cpu, mem, swap, netIn, netOut, tcpCount, udpCount, load1, online).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get metric history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric name (cpu, mem, swap, netIn, netOut, tcpCount, udpCount, load1, online)",
                        "name": "metric",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Bucket size in seconds (2, 30, 60, 120, 180, 300)",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/importDB": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restores the panel DB from an uploaded SQLite file. Panel restarts after restore. Destructive.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Import database",
                "parameters": [
                    {
                        "type": "file",
                        "description": "SQLite database file",
                        "name": "db",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/installXray/{version}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Downloads and installs the specified Xray version. Pass \"latest\" for the newest release.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Install/update Xray",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Xray tag (e.g. v25.10.31) or 'latest'",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.getDbPasswordRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
...
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Use format: \"Bearer \u003ctoken\u003e\". Get your token from Settings → Security → API Token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}